+++ airless/hook/file/file.py
@@ -3,9 +3,9 @@ import json
 import ndjson
 import os
 import requests
+import uuid
 
 from datetime import datetime
-from random import randint
 
 from airless.hook.base import BaseHook
 
@@ -32,7 +32,7 @@ class FileHook(BaseHook):
         filename = self.extract_filename(filepath_or_url)
         if add_timestamp:
             timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
-            filename = f'{timestamp}_{randint(1, 100000000)}_{filename}'
+            filename = f"{timestamp}_{uuid.uuid4().hex}_{filename}"
         return f'/tmp/{filename}'
 
     def download(self, url, headers, timeout=500, proxies=None):